requirements.txt
Django
django-filter==24.2
djangorestframework==3.15.1
djoser==2.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
ujson==5.9.0
celery
gunicorn
drf-spectacular
social-auth-app-django
django-jet-reboot
google-api-python-client==1.4.1
django-versatileimagefield==3.1
